fix(core-api): widen entry DTO types and stop rejecting zero-score saves
Entry.Valid() was silently discarding legitimate note-game results:

- UserID int16 -> int64: capped user ids at 32767.
- NPM int8 -> float64: capped notes-per-minute at 127, so a fast
  player's score above that failed JSON decode with a 400.
- Valid() no longer rejects CorrectQuestions == 0: a 0/10 game (e.g. a
  beginner who missed everything) was rejected as "missing" instead of
  saved.
- Valid() now rejects TotalQuestions <= 0, closing the gap left by the
  previous no-op presence check.
- NPM's presence rule is replaced with a negativity check (NPM < 0):
  the frontend computes npm = total/timeElapsed and rounds it, so a
  slow game can legitimately round down to 0 and must still save.

Refs #252

+++ core-api/generation/gen_utilities.go
@@ -336,8 +336,8 @@ func generateRealisticNoteGameEntries(studentID int16, entryCount int, profile S
 			TimeLength:       timeLength,
 			TotalQuestions:   int16(totalQuestions),
 			CorrectQuestions: int16(correctQuestions),
-			NPM:              int8(actualNPM),
-			UserID:           studentID,
+			NPM:              float64(actualNPM),
+			UserID:           int64(studentID),
 			CreatedDate: sql.NullString{
 				String: entryDate.Format("2006-01-02"),
 				Valid:  true,
